+++ board/views/__init__.py
@@ -1,3 +1,3 @@
 from .post_CRUD_views import *
-from .comment_CURD_views import *
+from .comment_CRUD_views import *
 from .post_list_view import *
